mcp-servers/capture/pyproject.toml
[project]
name = "cloto-mcp-capture"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "Pillow>=10.0.0",
    "mss>=9.0.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
